+++ examples/counter/app.ts
@@ -7,6 +7,9 @@ import {
   setProperties,
   state,
   when,
+  switchRender,
+  on,
+  onDefault,
 } from '@estrela';
 import { defer, map, startWith } from 'rxjs';
 
@@ -31,6 +34,18 @@ const Counter: Fel = () => {
     <ul>
       ${asyncMap(data, item => html`<li>${item}</li>`, html`<li>Loading...</li>`)}
     </ul>
+    <div>
+      ${switchRender(
+        count(),
+        on(0, html`<span>Zero</span>`),
+        on(1, html`<span>One</span>`),
+        on(2, html`<span>Two</span>`),
+        on(3, html`<span>Three</span>`),
+        on(4, html`<span>Four</span>`),
+        on(5, html`<span>Five</span>`),
+        onDefault(html`<span>Others</span>`)
+      )}
+    </div>
   `;
 };