+++ src/directives/index.ts
@@ -1,3 +1,4 @@
-export * from './async-render';
 export * from './async-map';
+export * from './async-render';
+export * from './switch-render';
 export * from './when';